+++ lib/src/controls/navigation/navigation_view/pane.dart
@@ -87,7 +87,10 @@ class NavigationPane with Diagnosticable {
     this.scrollController,
     this.leading,
     this.indicator = const StickyNavigationIndicator(),
-  }) : assert(selected == null || selected >= 0);
+  }) : assert(
+          selected == null || !selected.isNegative,
+          'The selected index must not be negative',
+        );
 
   final Key? key;
 
@@ -192,7 +195,7 @@ class NavigationPane with Diagnosticable {
 
   void _changeTo(NavigationPaneItem item) {
     final index = effectiveIndexOf(item);
-    if (selected != index) onChanged?.call(index);
+    if (selected != index && !index.isNegative) onChanged?.call(index);
   }
 
   /// A list of all of the items displayed on this pane.
@@ -208,6 +211,12 @@ class NavigationPane with Diagnosticable {
   /// Check if the provided [item] is selected on not.
   bool isSelected(NavigationPaneItem item) {
     return effectiveIndexOf(item) == selected;
+  }
+
+  /// Get the current selected item
+  PaneItem get selectedItem {
+    assert(selected != null, 'There is no item selected');
+    return effectiveItems[selected!] as PaneItem;
   }
 
   /// Get the effective index of the navigation pane.
@@ -383,7 +392,7 @@ abstract class NavigationPaneWidget {
 /// Creates a top navigation pane.
 ///
 /// ![Top Pane Anatomy](https://docs.microsoft.com/en-us/windows/uwp/design/controls-and-patterns/images/navview-pane-anatomy-horizontal.png)
-class _TopNavigationPane extends StatelessWidget {
+class _TopNavigationPane extends StatefulWidget {
   _TopNavigationPane({
     required this.pane,
     this.listKey,
@@ -394,85 +403,270 @@ class _TopNavigationPane extends StatelessWidget {
   final GlobalKey? listKey;
   final NavigationAppBar? appBar;
 
+  @override
+  State<_TopNavigationPane> createState() => _TopNavigationPaneState();
+}
+
+class _TopNavigationPaneState extends State<_TopNavigationPane> {
+  final overflowController = FlyoutController();
+  List<int> hiddenPaneItems = [];
+  late List<int> _localItemHold;
+  void generateLocalItemHold() {
+    _localItemHold = List.generate(
+      widget.pane.items.length,
+      (index) => index,
+    );
+  }
+
+  @override
+  void initState() {
+    super.initState();
+    generateLocalItemHold();
+  }
+
+  void _onPressed(PaneItem item) {
+    widget.pane._changeTo(item);
+  }
+
   Widget _buildItem(BuildContext context, NavigationPaneItem item) {
     if (item is PaneItemHeader) {
       return item.build(context);
     } else if (item is PaneItemSeparator) {
       return item.build(context, Axis.vertical);
     } else if (item is PaneItem) {
-      final selected = pane.isSelected(item);
+      final selected = widget.pane.isSelected(item);
       return item.build(
         context,
         selected,
-        () {
-          pane._changeTo(item);
-        },
-        showTextOnTop: !pane.footerItems.contains(item),
+        () => _onPressed(item),
+        // only show the text if the item is not in the footer
+        showTextOnTop: !widget.pane.footerItems.contains(item),
         displayMode: PaneDisplayMode.top,
       );
     } else {
       throw UnsupportedError(
-        '${item.runtimeType} is not a supported pane item type.',
+        '${item.runtimeType} is not a supported navigation pane item type.',
       );
     }
   }
 
   @override
+  void dispose() {
+    overflowController.dispose();
+    super.dispose();
+  }
+
+  @override
+  void didUpdateWidget(covariant _TopNavigationPane oldWidget) {
+    // update the items
+    if (oldWidget.pane.items.length != widget.pane.items.length) {
+      generateLocalItemHold();
+    }
+
+    // if the selected item changed
+    if (widget.pane.selected != oldWidget.pane.selected) {
+      final selectedItem = widget.pane.items.indexOf(
+        widget.pane.selectedItem,
+      );
+
+      // if the selected item is part of the middle items and
+      // if there is a non-hidden item
+      // and if the selected item is hidden
+      if (!selectedItem.isNegative &&
+          !hiddenPaneItems.contains(0) &&
+          hiddenPaneItems.contains(_localItemHold.indexOf(selectedItem))) {
+        generateLocalItemHold();
+
+        int item = hiddenPaneItems.first - 1;
+        while (widget.pane.items[item] is! PaneItem) {
+          item--;
+          if (item.isNegative) {
+            item++;
+            break;
+          }
+        }
+
+        _localItemHold
+          ..remove(selectedItem)
+          ..insert(item, selectedItem);
+        // print(
+        //   's$selectedItem to$item - i$_localItemHold - h$hiddenPaneItems',
+        // );
+      }
+    }
+
+    super.didUpdateWidget(oldWidget);
+  }
+
+  @override
   Widget build(BuildContext context) {
     assert(debugCheckHasFluentTheme(context));
-    Widget topBar = SizedBox(
-      height: pane.size?.topHeight ?? kOneLineTileHeight,
-      child: Row(key: pane.paneKey, children: [
+    final height = widget.pane.size?.topHeight ?? kOneLineTileHeight;
+    return SizedBox(
+      key: widget.pane.paneKey,
+      height: height,
+      child: Row(children: [
+        if (widget.pane.leading != null)
+          Padding(
+            padding: const EdgeInsets.symmetric(
+              horizontal: 8.0,
+              vertical: 6.0,
+            ),
+            child: widget.pane.leading!,
+          ),
+        if (widget.pane.header != null)
+          Padding(
+            padding: const EdgeInsets.symmetric(
+              horizontal: 8.0,
+              vertical: 6.0,
+            ),
+            child: widget.pane.header!,
+          ),
         Expanded(
-          child: Row(children: [
-            if (pane.leading != null)
-              Padding(
-                padding: const EdgeInsets.symmetric(
-                  horizontal: 8.0,
-                  vertical: 6.0,
-                ),
-                child: pane.leading!,
+          child: DynamicOverflow(
+            children: _localItemHold.map((index) {
+              final item = widget.pane.items[index];
+              return SizedBox(
+                height: height,
+                child: _buildItem(context, item),
+              );
+            }).toList(),
+            overflowWidgetAlignment: MainAxisAlignment.start,
+            overflowWidget: Flyout(
+              controller: overflowController,
+              child: PaneItem(icon: const Icon(FluentIcons.more)).build(
+                context,
+                false,
+                overflowController.open,
+                showTextOnTop: false,
+                displayMode: PaneDisplayMode.top,
               ),
-            if (pane.header != null)
-              Padding(
-                padding: const EdgeInsets.symmetric(
-                  horizontal: 8.0,
-                  vertical: 6.0,
-                ),
-                child: pane.header!,
-              ),
-            // TODO: A listview shouldn't be used here. Instead, if there are
-            // more items than space, show a dropdown button with the other
-            // items
-            Expanded(
-              child: SingleChildScrollView(
-                key: listKey,
-                primary: true,
-                scrollDirection: Axis.horizontal,
-                child: Row(
-                  children: pane.items.map((item) {
-                    return _buildItem(context, item);
-                  }).toList(),
-                ),
+              placement: FlyoutPlacement.end,
+              content: (context) => MenuFlyout(
+                items: _localItemHold.sublist(hiddenPaneItems.first).map((i) {
+                  final item = widget.pane.items[i];
+                  return buildMenuPaneItem(context, item);
+                }).toList(),
               ),
             ),
-          ]),
+            overflowChangedCallback: (hiddenItems) {
+              setState(() {
+                // indexes should always be valid
+                assert(() {
+                  for (var i = 0; i < hiddenItems.length; i++) {
+                    if (hiddenItems[i] < 0 ||
+                        hiddenItems[i] >= widget.pane.items.length) {
+                      return false;
+                    }
+                  }
+                  return true;
+                }());
+
+                hiddenPaneItems = hiddenItems;
+              });
+            },
+          ),
         ),
-        if (pane.autoSuggestBox != null)
+        if (widget.pane.autoSuggestBox != null)
           Container(
             margin: const EdgeInsets.only(left: 30.0),
             constraints: const BoxConstraints(
               minWidth: 100.0,
               maxWidth: 215.0,
             ),
-            child: pane.autoSuggestBox!,
+            child: widget.pane.autoSuggestBox!,
           ),
-        ...pane.footerItems.map((item) {
+        ...widget.pane.footerItems.map((item) {
           return _buildItem(context, item);
         }).toList(),
       ]),
     );
-    return topBar;
+  }
+
+  MenuFlyoutItemInterface buildMenuPaneItem(
+      BuildContext context, NavigationPaneItem item) {
+    if (item is PaneItemSeparator) {
+      return const MenuFlyoutSeparator();
+    } else if (item is PaneItem) {
+      return _MenuFlyoutPaneItem(
+        item: item,
+        onPressed: () => _onPressed(item),
+      );
+    } else {
+      throw UnsupportedError('${item.runtimeType} is not supported');
+    }
+  }
+}
+
+class _MenuFlyoutPaneItem extends MenuFlyoutItemInterface {
+  _MenuFlyoutPaneItem({
+    Key? key,
+    required this.item,
+    required this.onPressed,
+    this.selected = false,
+  }) : super(key: key);
+
+  final PaneItem item;
+  final VoidCallback? onPressed;
+  final bool selected;
+
+  @override
+  Widget build(BuildContext context) {
+    final size = PopupContentSizeInfo.of(context).size;
+    final NavigationPaneThemeData theme = NavigationPaneTheme.of(context);
+
+    final String titleText = item._getPropertyFromTitle<String>() ?? '';
+    final TextStyle baseStyle =
+        item._getPropertyFromTitle<TextStyle>() ?? const TextStyle();
+
+    final textResult = titleText.isNotEmpty
+        ? Padding(
+            padding: theme.labelPadding ?? EdgeInsets.zero,
+            child: RichText(
+              text: item._getPropertyFromTitle<InlineSpan>(baseStyle)!,
+              maxLines: 1,
+              overflow: TextOverflow.fade,
+              softWrap: false,
+              textAlign:
+                  item._getPropertyFromTitle<TextAlign>() ?? TextAlign.start,
+              textHeightBehavior:
+                  item._getPropertyFromTitle<TextHeightBehavior>(),
+              textWidthBasis: item._getPropertyFromTitle<TextWidthBasis>() ??
+                  TextWidthBasis.parent,
+            ),
+          )
+        : const SizedBox.shrink();
+
+    return HoverButton(
+      onPressed: onPressed,
+      builder: (context, states) {
+        return Container(
+          width: size.isEmpty ? null : size.width,
+          padding: MenuFlyout.itemsPadding,
+          height: 36.0,
+          color: ButtonThemeData.uncheckedInputColor(
+            FluentTheme.of(context),
+            states,
+          ),
+          child: Row(mainAxisSize: MainAxisSize.min, children: [
+            Padding(
+              padding: theme.iconPadding ?? EdgeInsets.zero,
+              child: IconTheme.merge(
+                data: IconThemeData(
+                  color: (selected
+                          ? theme.selectedIconColor?.resolve(states)
+                          : theme.unselectedIconColor?.resolve(states)) ??
+                      baseStyle.color,
+                  size: 16.0,
+                ),
+                child: Center(child: item.icon),
+              ),
+            ),
+            Flexible(child: textResult),
+          ]),
+        );
+      },
+    );
   }
 }