+++ lib/src/styles/theme.dart
@@ -326,7 +326,8 @@ class ThemeData with Diagnosticable {
     disabledColor ??=
         isLight ? const Color(0xFF838383) : Colors.grey[80].withOpacity(0.6);
     shadowColor ??= isLight ? Colors.black : Colors.grey[130];
-    scaffoldBackgroundColor ??= Colors.white.withOpacity(0.025);
+    scaffoldBackgroundColor ??=
+        isLight ? Colors.white : Colors.white.withOpacity(0.025);
     acrylicBackgroundColor ??= isLight
         ? const Color.fromARGB(204, 255, 255, 255)
         : const Color(0x7F1e1e1e);